Move all preprocessing to one file
